Code cleanup, generic pairables, keyboard features.
Removed redundant code, moved localisation and tooltip keys to a constants class, created generic pairable devices class this makes the keyboard better, and allows for future pairable devices.
localisation cleanup for future plans with the advanced printer.
keyboards no longer loos their connection when broken.

+++ src/main/java/com/theoriginalbit/minecraft/moarperipherals/tile/TileIronNote.java
@@ -45,8 +45,8 @@ public class TileIronNote extends TilePeripheral {
                 break;
         }
 
-        world.playSoundEffect((double) x + 0.5D, (double) y + 0.5D, (double) z + 0.5D, "note." + s, 3.0F, f);
-        world.spawnParticle("note", (double) x + 0.5D, (double) y + 1.2D, (double) z + 0.5D, (double) pitch / 24.0D, 0.0D, 0.0D);
+        world.playSoundEffect(x + 0.5D, y + 0.5D, z + 0.5D, "note." + s, 3.0F, f);
+        world.spawnParticle("note", x + 0.5D, y + 1.2D, z + 0.5D, (double) pitch / 24.0D, 0.0D, 0.0D);
     }
 
     @LuaFunction
